refactor

+++ src/main/java/com/seldon/mih/model/TermFrequenceIDF.java
@@ -3,7 +3,7 @@ package com.seldon.mih.model;
 public class TermFrequenceIDF {
     public String term;
     public String doc;
-    public Integer count;
+    public Long count;
     public Double tf;
     public Double tfIDF;
 
@@ -11,7 +11,7 @@ public class TermFrequenceIDF {
 
     }
 
-    public TermFrequenceIDF(String term, String doc, Integer count, Double tf, Double tfIDF) {
+    public TermFrequenceIDF(String term, String doc, Long count, Double tf, Double tfIDF) {
         this.term = term;
         this.doc = doc;
         this.count = count;
